@@ -1,50 +1,45 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrash, faFileDownload } from "@fortawesome/free-solid-svg-icons";
-import { library } from "@fortawesome/fontawesome-svg-core"; //allows later to just use icon name to render-them
 
-library.add(faTrash, faFileDownload);
+function ResetAndSave({ handleReset, handleSave, isDark = false }) {
+  const [hoveredButton, setHoveredButton] = useState(null);
 
-class ResetAndSave extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
+  const buttonStyle = (isHovered) => ({
+    display: 'flex',
+    alignItems: 'center',
+    padding: '8px 12px',
+    fontSize: '13px',
+    color: isHovered ? 'var(--text-primary)' : 'var(--text-secondary)',
+    backgroundColor: isHovered ? 'var(--bg-tertiary)' : 'transparent',
+    border: 'none',
+    borderRadius: '8px',
+    cursor: 'pointer',
+    transition: 'all 0.15s ease',
+  });
 
-  render() {
-    return (
-      <div className="absolute inline-block ml-4 mt-7">
-        <div className="inline-block list-none">
-          <li
-            key="Reset"
-            onClick={this.props.handleReset}
-            className="px-3 py-1 mt-2 text-base text-gray-500 rounded-lg hover:text-gray-900 hover:cursor-pointer"
-          >
-            <FontAwesomeIcon
-              icon={["fas", "trash"]}
-              size="lg"
-              className="mx-2 -mt-0.5"
-            />
-            Reset
-          </li>
-        </div>
-        <div className="inline-block list-none">
-          <li
-            key="Save"
-            onClick={this.props.handleSave}
-            className="px-3 py-1 mt-2 text-base text-gray-500 rounded-lg hover:text-gray-900 hover:cursor-pointer"
-          >
-            <FontAwesomeIcon
-              icon={["fas", "file-download"]}
-              size="lg"
-              className="mx-2 -mt-0.5"
-            />
-            Save
-          </li>
-        </div>
-      </div>
-    );
-  }
+  return (
+    <div style={{ display: 'flex', alignItems: 'center', gap: '4px' }}>
+      <button
+        onClick={handleReset}
+        style={buttonStyle(hoveredButton === 'reset')}
+        onMouseEnter={() => setHoveredButton('reset')}
+        onMouseLeave={() => setHoveredButton(null)}
+      >
+        <FontAwesomeIcon icon={faTrash} style={{ marginRight: 8 }} />
+        Reset
+      </button>
+      <button
+        onClick={handleSave}
+        style={buttonStyle(hoveredButton === 'save')}
+        onMouseEnter={() => setHoveredButton('save')}
+        onMouseLeave={() => setHoveredButton(null)}
+      >
+        <FontAwesomeIcon icon={faFileDownload} style={{ marginRight: 8 }} />
+        Save
+      </button>
+    </div>
+  );
 }
 
 export default ResetAndSave;
